+++ package.json
@@ -16,6 +16,7 @@
     "meshline": "^3.1.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.12.1",
     "react-scripts": "5.0.1",
     "react-three-fiber": "^6.0.13",
     "styled-components": "^6.0.0-rc.3",
