+++ src/pages/NoMatch.js
@@ -1,0 +1,5 @@
+function NoMatch() {
+    return <h2>このページは存在しません。</h2>;
+  }
+  
+  export default NoMatch;
